+++ dandy/intel/factory.py
@@ -13,10 +13,10 @@ class IntelFactory:
 
     @staticmethod
     def _run_for_intel_class_or_object(
-            intel: BaseIntel | Type[BaseIntel],
-            class_func: Callable,
-            object_func: Callable,
-            **kwargs,
+        intel: BaseIntel | Type[BaseIntel],
+        class_func: Callable,
+        object_func: Callable,
+        **kwargs,
     ) -> Any:
         if isinstance(intel, BaseIntel):
             return object_func(**kwargs)
@@ -28,30 +28,37 @@ class IntelFactory:
 
     @classmethod
     def intel_to_json_inc_ex_schema(
-            cls,
-            intel: BaseIntel | Type[BaseIntel],
-            include: IncEx | None = None,
-            exclude: IncEx | None = None,
+        cls,
+        intel: BaseIntel | Type[BaseIntel],
+        include: IncEx | None = None,
+        exclude: IncEx | None = None,
     ) -> Dict:
         inc_ex_kwargs = {'include': include, 'exclude': exclude}
 
-        return cls._run_for_intel_class_or_object(
+        json_schema = cls._run_for_intel_class_or_object(
             intel=intel,
             class_func=intel.model_json_inc_ex_schema,
             object_func=intel.model_object_json_inc_ex_schema,
-            **inc_ex_kwargs
+            **inc_ex_kwargs,
         )
+
+        for property_ in json_schema['properties']:
+            for required_value in (
+                'type',
+            ):
+                if required_value not in json_schema['properties'][property_]:
+                    message = f'JSON Schema property {property_} is missing required value {required_value}'
+                    raise IntelCriticalException(message) from None
+
+        return json_schema
 
     @classmethod
     def json_str_to_intel_object(
-            cls,
-            json_str: str,
-            intel: BaseIntel | Type[BaseIntel]
+        cls, json_str: str, intel: BaseIntel | Type[BaseIntel]
     ) -> BaseIntel:
-
         return cls._run_for_intel_class_or_object(
             intel=intel,
             class_func=intel.model_validate_json,
             object_func=intel.model_validate_json_and_copy,
-            json_data=json_str
+            json_data=json_str,
         )